// app-ui/src/app/backend/model/analysisData.ts
/**
 * TIBCO Discover public API
 * TIBCO Discover public API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TypeValue } from './typeValue';
import { Mapping } from './mapping';


export interface AnalysisData { 
    /**
     * TBD
     */
    name: string;
    /**
     * TBD
     */
    description: string;
    /**
     * TBD
     */
    datasetId: string;
    /**
     * TBD
     */
    templateId?: string;
    /**
     * Template label
     */
    templateLabel?: string;
    mappings: Mapping;
    filters?: Array<TypeValue>;
    groups?: Array<TypeValue>;
    /**
     * TBD
     */
    progress?: number;
}